refactor(frontend): split Layout shell
