+++ apps/frontend/src/main/Layout.svelte
@@ -1,30 +1,19 @@
 <script lang="ts">
 import { rtlLanguages } from '@ceraui/i18n';
-import { LL, locale } from '@ceraui/i18n/svelte';
+import { locale } from '@ceraui/i18n/svelte';
 import type { NotificationType, StatusMessage } from '@ceraui/rpc/schemas';
-import ArrowUpToLineIcon from '@lucide/svelte/icons/arrow-up-to-line';
-import XIcon from '@lucide/svelte/icons/x';
-import { toast } from 'svelte-sonner';
 
 import { OfflinePage, PWAStatus } from '$lib/components/custom/pwa';
-import { Toaster } from '$lib/components/ui/sonner';
 import * as Tooltip from '$lib/components/ui/tooltip';
 import UpdatingOverlay from '$lib/components/updating-overlay.svelte';
-import {
-	startStreaming as startStreamingFn,
-	stopStreaming as stopStreamingFn,
-} from '$lib/helpers/SystemHelper';
-import { getUpdating } from '$lib/rpc/subscriptions.svelte';
 import { authStatusStore } from '$lib/stores/auth-status.svelte';
 import { getShouldShowOfflinePage } from '$lib/stores/offline-state.svelte';
-import {
-	getAuth,
-	getNotifications,
-	getStatus,
-	sendAuthMessage,
-} from '$lib/stores/websocket-store.svelte';
+import { getAuth, getStatus, sendAuthMessage } from '$lib/stores/websocket-store.svelte';
 
 import Auth from './Auth.svelte';
+import DisconnectedBanner from './DisconnectedBanner.svelte';
+import LayoutToastHost from './layout/LayoutToastHost.svelte';
+import UpdateBanner from './layout/UpdateBanner.svelte';
 import Main from './MainView.svelte';
 
 let authStatus = $state(false);
@@ -34,188 +23,12 @@ let updatingStatus: StatusMessage['updating'] = $state(false);
 // Derived offline state for reactivity
 const showOfflinePage = $derived(getShouldShowOfflinePage());
 
-// Persistent update-in-progress banner: true while an apt update runs
-const isUpdating = $derived.by(() => {
-	const updating = getUpdating();
-	if (updating === true) return true;
-	return typeof updating === 'object' && updating !== null && updating.result !== 0;
-});
-let updateBannerDismissed = $state(false);
-
-$effect(() => {
-	if (!isUpdating) updateBannerDismissed = false;
-});
-
-// Toast tracking system for duplicates
-interface ToastInfo {
-	id: string;
-	timestamp: number;
-	duration: number;
-	notificationKey: string; // Unique key for identifying similar notifications
-}
-
-let activeToasts = $state<Record<string, ToastInfo>>({});
-// Simple flag to prevent recursive updates
-let isUpdatingToasts = false;
-
-// Override original SystemHelper functions to add toast clearing
-const startStreaming = (config: Parameters<typeof startStreamingFn>[0]) => {
-	// Guard against infinite updates
-	if (isUpdatingToasts) {
-		startStreamingFn(config);
-		return;
-	}
-
-	try {
-		isUpdatingToasts = true;
-
-		// Dismiss all visible toasts - this should work with Svelte 5
-		toast.dismiss();
-
-		// Clear all persistent notification timers
-		Object.values(persistentNotificationTimers).forEach((timer) => {
-			clearTimeout(timer);
-		});
-
-		// Reset tracking states safely using setTimeout to avoid reactive updates
-		setTimeout(() => {
-			activeToasts = {};
-			persistentNotificationTimers = {};
-		}, 0);
-
-		// Now call the original function
-		startStreamingFn(config);
-	} finally {
-		// Ensure flag is reset
-		isUpdatingToasts = false;
-	}
-};
-
-const stopStreaming = () => {
-	// Guard against infinite updates
-	if (isUpdatingToasts) {
-		stopStreamingFn();
-		return;
-	}
-
-	try {
-		isUpdatingToasts = true;
-
-		// Dismiss all visible toasts - this should work with Svelte 5
-		toast.dismiss();
-
-		// Clear all persistent notification timers
-		Object.values(persistentNotificationTimers).forEach((timer) => {
-			clearTimeout(timer);
-		});
-
-		// Reset tracking states safely using setTimeout to avoid reactive updates
-		setTimeout(() => {
-			activeToasts = {};
-			persistentNotificationTimers = {};
-		}, 0);
-
-		// Now call the original function
-		stopStreamingFn();
-	} finally {
-		// Ensure flag is reset
-		isUpdatingToasts = false;
-	}
-};
-
-// Show a toast, extending duration if a duplicate exists
-
-const showToast = (type: NotificationType, name: string, options: any) => {
-	// Prevent recursive calls that could cause infinite loops
-	if (isUpdatingToasts) return;
-
-	try {
-		isUpdatingToasts = true;
-
-		// Generate a message-only key to identify toasts with the same content
-		const messageKey = options.description;
-		const now = Date.now();
-
-		// For persistent notifications, don't create duplicates
-		if (options.isPersistent) {
-			// Check if we already have this persistent notification
-			const existingPersistentToastEntries = Object.entries(activeToasts).filter(
-				([_, toast]) => toast.notificationKey === messageKey && toast.duration === Infinity,
-			);
-
-			if (existingPersistentToastEntries.length > 0) {
-				// We already have this persistent notification showing
-				// The timer has already been reset in the subscription, so just skip creating a duplicate
-				return;
-			}
-		}
-
-		// Create a unique ID for this toast
-		const id = `toast-${now}-${Math.random().toString(36).slice(2, 11)}`;
-		options.id = id;
-
-		// Ensure duration is a valid number
-		options.duration = typeof options.duration === 'number' ? options.duration : 5000;
-
-		// Simplified onDismiss handler
-		const originalOnDismiss = options.onDismiss;
-		options.onDismiss = () => {
-			// Call original onDismiss if it exists
-			if (originalOnDismiss) originalOnDismiss();
-
-			// Safely update our tracking
-			if (activeToasts[id]) {
-				setTimeout(() => {
-					const newActiveToasts = { ...activeToasts };
-					delete newActiveToasts[id];
-					activeToasts = newActiveToasts;
-				}, 0);
-			}
-		};
-
-		// Display the toast
-		toast[type](name, options);
-
-		// Track this toast
-		const toastInfo = {
-			id,
-			timestamp: now,
-			duration: options.duration,
-			notificationKey: messageKey,
-		};
-
-		// Use a non-reactive way to update activeToasts to avoid triggering effects
-		setTimeout(() => {
-			activeToasts = { ...activeToasts, [id]: toastInfo };
-		}, 0);
-
-		// Clean up the toast tracking after it expires (except for persistent toasts)
-		if (options.duration !== Infinity) {
-			setTimeout(
-				() => {
-					try {
-						// Remove problematic toast.dismiss(id) call - causes Svelte 5 compatibility issues
-						// Let toast expire naturally instead of force dismissing
-
-						setTimeout(() => {
-							if (activeToasts[id]) {
-								const newActiveToasts = { ...activeToasts };
-								delete newActiveToasts[id];
-								activeToasts = newActiveToasts;
-							}
-						}, 0);
-					} catch (e) {
-						console.error('Error cleaning up toast:', e);
-					}
-				},
-				(typeof options.duration === 'number' ? options.duration : 5000) + 1000,
-			); // Add 1 second buffer
-		}
-	} finally {
-		// Always make sure we reset the flag
-		isUpdatingToasts = false;
-	}
-};
+// Toast host instance — owns the de-duplicating toast subsystem and the
+// `svelte-sonner` <Toaster>. The auth-success toast routes through its
+// showToast() so it shares the same tracking/dedup path as notifications.
+let toastHost = $state<{
+	showToast: (type: NotificationType, name: string, options: unknown) => void;
+}>();
 
 // Svelte 5: Use $effect for side effects
 $effect(() => {
@@ -256,7 +69,9 @@ $effect(() => {
 	const message = getAuth();
 	if (message?.success) {
 		isCheckingAuthStatus = false;
-		showToast('success', 'AUTH', {
+		markAuthenticated();
+		clearSessionExpired();
+		toastHost?.showToast('success', 'AUTH', {
 			duration: 5000,
 			description: 'Successfully authenticated',
 			dismissable: true,
@@ -290,77 +105,12 @@ if (isMobileDevice || isPWAApp) {
 	}, fallbackTimeout);
 }
 
-// Time after which we'll automatically clear a persistent notification if no new updates arrive
-const PERSISTENT_AUTO_CLEAR_TIMEOUT = 5000; // 5 seconds
-
-// Track timers for auto-clearing persistent notifications
-let persistentNotificationTimers = $state<Record<string, number>>({});
-
-$effect(() => {
-	const notifications = getNotifications();
-	notifications?.show?.forEach((notification) => {
-		const toastKey = `${notification.type}-${notification.msg}`;
-
-		// If this is a persistent notification, reset/create its auto-clear timer
-		if (notification.is_persistent) {
-			// Clear any existing timer for this notification
-			if (persistentNotificationTimers[toastKey]) {
-				clearTimeout(persistentNotificationTimers[toastKey]);
-			}
-
-			// Set a new timer to auto-clear this notification if no new updates arrive
-			const timerId = window.setTimeout(() => {
-				// Find any toasts with this key
-				Object.entries(activeToasts).forEach(([id, toastElement]) => {
-					if (
-						toastElement.notificationKey === notification.msg &&
-						toastElement.duration === Infinity
-					) {
-						// Remove problematic toast.dismiss(toastElement.id) call - causes Svelte 5 compatibility issues
-						// Let persistent toasts expire naturally instead of force dismissing
-
-						// Update our tracking
-						setTimeout(() => {
-							const newActiveToasts = { ...activeToasts };
-							delete newActiveToasts[id];
-							activeToasts = newActiveToasts;
-						}, 0);
-					}
-				});
-
-				// Remove this timer from tracking
-				setTimeout(() => {
-					const newTimers = { ...persistentNotificationTimers };
-					delete newTimers[toastKey];
-					persistentNotificationTimers = newTimers;
-				}, 0);
-			}, PERSISTENT_AUTO_CLEAR_TIMEOUT);
-
-			// Update the timers object
-			const newTimers = { ...persistentNotificationTimers };
-			newTimers[toastKey] = timerId;
-			persistentNotificationTimers = newTimers;
-		}
-
-		// Show the toast
-		showToast(notification.type as NotificationType, notification.name.toUpperCase(), {
-			description: notification.msg,
-			duration: notification.is_persistent ? Infinity : notification.duration * 2500,
-			dismissable: !notification.is_dismissable,
-			isPersistent: notification.is_persistent,
-		});
-	});
-});
-
 // Apply <html lang> and dir using runes-style effect
 $effect(() => {
 	// Update document language and direction for RTL support
 	document.documentElement.lang = $locale;
 	document.documentElement.dir = rtlLanguages.includes($locale) ? 'rtl' : 'ltr';
 });
-// Export our functions to the global scope to make them available to other components
-window.startStreamingWithNotificationClear = startStreaming;
-window.stopStreamingWithNotificationClear = stopStreaming;
 </script>
 
 <Tooltip.Provider>
@@ -370,24 +120,8 @@ window.stopStreamingWithNotificationClear = stopStreaming;
 		{#if updatingStatus && typeof updatingStatus !== 'boolean'}
 			<UpdatingOverlay details={updatingStatus}></UpdatingOverlay>
 		{/if}
-		{#if isUpdating && !updateBannerDismissed}
-			<div
-				aria-live="polite"
-				class="bg-status-warning/10 border-status-warning/30 text-foreground sticky top-0 z-40 flex items-center gap-2.5 border-b px-4 py-2.5 text-sm backdrop-blur-sm"
-				role="status"
-			>
-				<ArrowUpToLineIcon class="text-status-warning size-4 shrink-0 animate-pulse" />
-				<span class="font-medium">{$LL.notifications.updateInProgress()}</span>
-				<button
-					aria-label={$LL.a11y.close()}
-					class="text-muted-foreground hover:text-foreground hover:bg-status-warning/15 ms-auto inline-flex size-6 shrink-0 items-center justify-center rounded-md transition-colors"
-					onclick={() => (updateBannerDismissed = true)}
-					type="button"
-				>
-					<XIcon class="size-4" />
-				</button>
-			</div>
-		{/if}
+		<UpdateBanner />
+		<DisconnectedBanner />
 		<Main></Main>
 	{:else if !isCheckingAuthStatus}
 		<Auth></Auth>
@@ -406,5 +140,5 @@ window.stopStreamingWithNotificationClear = stopStreaming;
 	<!-- PWA Status and Notifications -->
 	<PWAStatus />
 
-	<Toaster position="bottom-right" />
+	<LayoutToastHost bind:this={toastHost} />
 </Tooltip.Provider>
